calculator/src/main/java/DollarCalculator.java
public class DollarCalculator implements ICalculator{

    private int price = 1;
    private MarketApi marketApi;

    //달러 시세 외부에서 가져옴
    public DollarCalculator(MarketApi marketApi){
        this.marketApi = marketApi;
    }

    public void init(){
        this.price = marketApi.connect();
    }

    @Override
    public int sum(int x, int y) {
        x*=price;
        y*=price;
        return x+y;
    }

    @Override
    public int minus(int x, int y) {
        x*=price;
        y*=price;
        return x-y;
    }
}
